requirements.txt
requests
discord.py
pillow
python-dotenv
pyinstaller
pyyaml
pyautogui
opencv-python
simpleaudio
pydub